+++ src/Console/Infrastructure/Command/ListModulesCommand.php
@@ -8,9 +8,11 @@ use Gacela\Console\ConsoleFacade;
 use Gacela\Console\Domain\AllAppModules\AppModule;
 use Gacela\Framework\DocBlockResolverAwareTrait;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -20,60 +22,77 @@ final class ListModulesCommand extends Command
 {
     use DocBlockResolverAwareTrait;
 
+    private const CHECK_SYMBOL = '✔️';
+    private const CROSS_SYMBOL = '✖️';
+
+    private ?OutputInterface $output = null;
+
     protected function configure(): void
     {
         $this->setName('list:modules')
             ->setDescription('Render all modules found')
             ->addArgument('filter', InputArgument::OPTIONAL, 'Any filter to simplify the output')
-            ->addOption('simple', 's', InputOption::VALUE_NONE, 'Display just the module names');
+            ->addOption('detailed', 'd', InputOption::VALUE_NONE, 'Display all the modules in detail');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $this->output = $output;
+
         $filter = (string)$input->getArgument('filter');
 
-        $listOfModules = $this->generateListOfModules(
-            (bool)$input->getOption('simple'),
+        $this->generateListOfModules(
+            (bool)$input->getOption('detailed'),
             $this->getFacade()->findAllAppModules($filter),
         );
-
-        $output->write($listOfModules);
 
         return self::SUCCESS;
     }
 
-    /**
-     * @param list<AppModule> $modules
-     */
-    private function generateListOfModules(bool $isSimple, array $modules): string
+    private function output(): OutputInterface
     {
-        return ($isSimple)
-            ? $this->generateSimpleView($modules)
-            : $this->generateDetailedView($modules);
+        return $this->output ?? new ConsoleOutput();
     }
 
     /**
      * @param list<AppModule> $modules
      */
-    private function generateSimpleView(array $modules): string
+    private function generateListOfModules(bool $isDetailed, array $modules): void
     {
-        $result = '';
-
-        foreach ($modules as $i => $module) {
-            $n = $i + 1;
-            $result .= <<<TXT
-{$n}.- <fg=green>{$module->moduleName()}</>
-
-TXT;
+        if ($isDetailed) {
+            $this->generateDetailedView($modules);
+        } else {
+            $this->generateSimpleView($modules);
         }
-
-        return $result;
     }
 
     /**
      * @param list<AppModule> $modules
      */
-    private function generateDetailedView(array $modules): string
+    private function generateSimpleView(array $modules): void
+    {
+        $rows = [];
+
+        foreach ($modules as $module) {
+            $rows[] = [
+                $module->fullModuleName(),
+                self::CHECK_SYMBOL, // facade is always true
+                $module->factoryClass() ? self::CHECK_SYMBOL : self::CROSS_SYMBOL,
+                $module->configClass() ? self::CHECK_SYMBOL : self::CROSS_SYMBOL,
+                $module->dependencyProviderClass() ? self::CHECK_SYMBOL : self::CROSS_SYMBOL,
+            ];
+        }
+        $table = new Table($this->output());
+        $table->setStyle('box');
+        $table->setHeaders(['Module namespace', 'Facade', 'Factory', 'Config', 'Dep. Provider']);
+        $table->setRows($rows);
+        $table->render();
+    }
+
+    /**
+     * @param list<AppModule> $modules
+     */
+    private function generateDetailedView(array $modules): void
     {
         $result = '';
         foreach ($modules as $i => $module) {
@@ -93,6 +112,7 @@ TXT;
 
 TXT;
         }
-        return $result;
+
+        $this->output()->write($result);
     }
 }